// gen/workout/v1/workout.schema.ts
export interface components {
    schemas: {
        WorkoutServiceFinishWorkoutBody: Record<string, never>;
        protobufAny: {
            "@type"?: string;
        } & {
            [key: string]: unknown;
        };
        rpcStatus: {
            /** Format: int32 */
            code?: number;
            message?: string;
            details?: components["schemas"]["protobufAny"][];
        };
        v1FinishWorkoutResponse: {
            /** Format: date-time */
            finishedAt?: string;
        };
        v1GetWorkoutResponse: {
            workout?: components["schemas"]["v1Workout"];
            sets?: components["schemas"]["v1Set"][];
        };
        v1ListWorkoutsResponse: {
            workouts?: components["schemas"]["v1Workout"][];
        };
        v1Set: {
            setId?: string;
            workoutId?: string;
            exerciseId?: string;
            /** Format: int32 */
            rep?: number;
            /** Format: double */
            weight?: number;
            /** Format: date-time */
            trainedAt?: string;
            /** Format: date-time */
            createdAt?: string;
        };
        v1StartWorkoutRequest: Record<string, never>;
        v1StartWorkoutResponse: {
            workoutId?: string;
        };
        v1Workout: {
            workoutId?: string;
            /** Format: date-time */
            startedAt?: string;
            /** Format: date-time */
            finishedAt?: string;
            /** Format: date-time */
            createdAt?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}